Change InputState variable types to float because of SDL3 transition

+++ src/io/InputState.hpp
@@ -9,15 +9,15 @@ struct InputState
     std::bitset<SDL_SCANCODE_COUNT> released;  // buttons that transitioned 1->0 this frame
 
     // Mouse
-    int mouseX = 0; // absolute motion position
-    int mouseY = 0;               
-    int dx = 0; // total motion during this frame
-    int dy = 0;
+    float mouseX = 0; // absolute motion position
+    float mouseY = 0;               
+    float dx = 0; // total motion during this frame
+    float dy = 0;
     uint32_t mouseDown = 0; // SDL_BUTTON_LEFT, SDL_BUTTON_RIGHT or SDL_BUTTON_MIDDLE
     uint32_t mousePressed = 0;
     uint32_t mouseReleased = 0;
-    int wheelX = 0;
-    int wheelY = 0;
+    float wheelX = 0;
+    float wheelY = 0;
 
     void Reset() 
     {
